Classification/requirements.txt
ipykernel==6.29.5
pandas==2.2.2
numpy==1.26.4
seaborn==0.13.2
matplotlib==3.9.2
scikit-learn==1.5.1
gdown==5.1.0
python-box==7.1.1
PyYAML==6.0.1
xgboost==2.1.1
streamlit==1.38.0
-e. 